+++ src/band3/meta_band/NetSync.cpp
@@ -12,7 +12,6 @@
 #include "obj/Object.h"
 #include "os/Debug.h"
 #include "ui/UI.h"
-#include "ui/UIMessages.h"
 #include "ui/UIPanel.h"
 #include "ui/UIScreen.h"
 #include "utl/Symbol.h"
