+++ src/band3/meta_band/OvershellSlot.h
@@ -8,7 +8,6 @@
 #include "bandobj/BandLabel.h"
 #include "meta_band/OvershellProfileProvider.h"
 #include "game/GameMessages.h"
-#include "ui/UIMessages.h"
 #include "os/JoypadMsgs.h"
 #include "os/VirtualKeyboard.h"
 #include "tour/TourCharLocal.h"
