+++ src/band3/meta_band/UIStats.h
@@ -3,7 +3,7 @@
 #include "obj/Object.h"
 #include "os/JoypadMsgs.h"
 #include "os/OnlineID.h"
-#include "ui/UIMessages.h"
+#include "ui/UIComponent.h"
 #include "ui/UIScreen.h"
 #include "utl/Str.h"
