+++ src/system/ui/UITrigger.cpp
@@ -1,7 +1,6 @@
 #include "ui/UITrigger.h"
 #include "ui/UIComponent.h"
 #include "obj/Task.h"
-#include "ui/UIMessages.h"
 #include "utl/Symbols.h"
 // #include "utl/ClassSymbols.h"
